# DDF-2474 Fixes issue with ingesting a large zip file with more metacards than space in blocking queue.