#### What does this PR do?
Fixes bug where some records might not be ingested with no error logged if the zip file to be ingested contained more than `--batchsize` * `--multithreaded` metacards.

This also cleans up the threading logic, relying on a Phaser used as a CountUp/DownLatch instead of flags and Thread.sleeps().

This needs to be backported to 2.9.x but might require a slightly more complex cherrypick as some other changes have been introduced to master and not backported.

#### Who is reviewing it (please choose AT LEAST two reviewers that need to approve the PR before it can get merged; if a component team is listed, at least one of its members needs to approve)?
@bdeining @rzwiefel @tbatie 

#### Choose 2 committers to review/merge the PR (please choose ONLY two committers from below, delete the rest).
@kcwire
@pklinef

#### How should this be tested?
- Full build/test run.
- Threading tests
    - Get some metacards to be ingested by any transformer and put them into a directory. I would recommend at least 100.
    - Ingest through the karaf command line (include the `-p` flag to speed the tests, passing various values for the `-b` (`--batchsize`) and `-m` (`--multithreaded`) flags to test different possible edge cases. e.g.
        - `-b 1 -m 1` 
        - `-b 1 -m 20` 
        - `-b [number of metacards in test]`
        - `-b [number of metacards in test] -m 1`
        - `-b [number of metacards in test] -m 3`
        - `-b [factor of metacard count]` (For 100, using 10 or 20 would be good for this)
        - `-b [factor of metacard count - 1]`
        - `-b [factor of metacard count + 1]`
        - `-b [metacard count/2]`
        - `-b [metacard count/2 - 1]`
        - `-b [metacard count/2 + 1]`
        - `-b [metacard count - 1]`
        - `-b [metacard count + 1]`
        - `-b [prime number less than metacard count]`
    - Run each of the above tests several times to ensure there are no timing issues. There is no need to remove the records from the catalog between tests. If the threading is going to fail, it's going to do so whether the records are already present of not.
- Zip ingest test
    - Create zip file of metacards to ingest, either manually or through `catalog:dump`
    - Sign the zip file. This can be accomplished by running `jarsigner -keystore $DDF_HOME/etc/keystores/serverKeystore.jks [zipfile] localhost`, assuming your DDF instance is named `localhost`.
    - Ingest the zip file, first with default batch and thread size, then by adjusting those parameters so their product is smaller than the number of metacards in the zip file. e.g.
        - `ingest --include-content [zipfile]`
        - `ingest --include-content -b 2 -m 2 [zipfile]` (If you have more than four metacards in your zip, these values would trigger the error prior to this fix.)

#### Any background context you want to provide?
N/A

#### What are the relevant tickets?
DDF-2474

#### Screenshots (if appropriate)
N/A

#### Checklist:
- [ ] Documentation Updated
- [ ] Update / Add Unit Tests
- [ ] Update / Add Integration Tests