@coyotesqrl 2016-09-12
@stustison asked for some metrics on any performance ramifications for the threading changes. I ran ingests using three different data sets against current master and this branch. The differences in performance were negligible:

Dataset Size | Import Transformer | Master Time to Ingest | Updated Threading
-------------| ------------------ | --------------------- | -----------------
21578        | xml                | 1:02                  | 1:16
34475        | xml                | 5:04                  | 4:53
1000000      | geojson            | 1:23:05               | 1:20:04