@pklinef 2016-09-13
I think this command will fail if the system cannot keep up with the queue.  The phaser only supports up to 65535 parties without tiering the phasers.

>Implementation notes: This implementation restricts the maximum number of parties to 65535. Attempts to register additional parties result in IllegalStateException. However, you can and should create tiered phasers to accommodate arbitrarily large sets of participants.
http://docs.oracle.com/javase/7/docs/api/java/util/concurrent/Phaser.html#bulkRegister(int)