@coyotesqrl 2016-09-15
Changes pushed in a separate commit. It might be unnecessary with the updated github review to keep it separate but just in case it isn't, I wanted this to be quick to look at.