@bdeining 2016-09-15
Full Hero build successful.  Ran through all suggested test procedures and observed expected ingest behavior throughout.  LGTM